ArmPkg/ArmGicDxe: Fix initialization of IPRIORITYR registers
This fixes the RAS GIC error detected during initialization of
the GICv3.

    Error: GIC - ERR0 {MUE}
    FR: 0x00000000000002A1
    CTLR: 0x0000000000000000
    STATUS: 0x000000006C20030F
    MISC0: 0x0000000000000000
    MISC1: 0x0000000000000000

Currently the ArmGic code tries to initialize all the IPRIORITYR
registers to a default state via the Distributor (GICD). However,
when Affinity Routing is Enabled, the first eight IPRIORITYR
registers, that control first 32 interrupts (SGI and PPI), do not
exist in GICD and can only be accessed via the Redistributor (GICR).

To access the Redistributor it needs to be powered up and initialized.
The Redistributor is powered up and initialized at the appropriate
time by the trusted firmware in gicv3_rdistif_init for each processor.
Because of this, we can skip initializing the first eight IPRIORITYR
registers as the trusted firmware has already done this step.

Signed-off-by: Victor Gallardo <Victor.Gallardo@amperecomputing.com>
Signed-off-by: Quan Nguyen <qnguyen@amperecomputing.com>