MdeModulePkg/UefiBootManagerLib: Add filter to remove invalid boot options
This patch adds a condition to stop the disks which are not removable
and does not have EFI System Partition to create boot options.

Currently, BootManagerLib will enumerate the boot option for block IO
which does not have EFI System Partition. For a bootable disk, it must
have EFI System Partition to store the boot loaders or kernel images.
Creating a boot option for block IO, which is not removable and does
not have ESP, should not happen.

Signed-off-by: Tinh Nguyen <tinhn@amperecomputing.com>