MdeModulePkg: Variable: Fix variable dupplication after NVRAM is full
Currently, VariableDxe maintains a cache of variables copied from the
NVRAM FV. Any variable updates must walk through this region to check
if the variable has existed or not. When NVRAM is almost Full, It will
trigger the ReClaim process, and the cache becomes the temporary buffer
to store the new variable layout. The problem is that at the end of the
ReClaim process, the cache will be written back to the old data. This
will lead to duplication if there is a modification to an existed
variable.

Signed-off-by: Vu Nguyen <Vu@amperecomputing.com>