ArmPkg: Add new library to modify DriverOrder/BootOrder variable
This library provides functions to manipulate the DriverOrder/BootOrder
variable in UEFI firmware for Arm platform. It can be used by applications
or drivers that need to change the boot order programmatically.

This library is recommended to use in PlatformBootmanagerLib
because it simplifies the boot order management and avoids conflicts
with other components.

Signed-off-by: Tinh Nguyen <tinhnguyen@os.amperecomputing.com>